+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/component/EntityComponentsTest.java
@@ -26,7 +26,6 @@ import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
-import is.codion.swing.common.model.component.combobox.ItemComboBoxModel;
 import is.codion.swing.common.ui.component.button.NullableCheckBox;
 import is.codion.swing.common.ui.component.combobox.Completion;
 import is.codion.swing.common.ui.component.text.TextFieldPanel;
@@ -140,7 +139,7 @@ public final class EntityComponentsTest {
 										.transferFocusOnEnter(true)
 										.link(editModel.value(Detail.BOOLEAN))
 										.buildValue();
-		ItemComboBoxModel<Boolean> boxModel = (ItemComboBoxModel<Boolean>) componentValue.component().getModel();
+		FilterComboBoxModel<Item<Boolean>> boxModel = (FilterComboBoxModel<Item<Boolean>>) componentValue.component().getModel();
 		assertTrue(boxModel.selection().value().value());
 		boxModel.setSelectedItem(null);
 		assertNull(editModel.value(Detail.BOOLEAN).get());
@@ -158,8 +157,8 @@ public final class EntityComponentsTest {
 										.buildValue();
 		JComboBox<Item<Integer>> comboBox = componentValue.component();
 
-		ItemComboBoxModel<Integer> model = (ItemComboBoxModel<Integer>) comboBox.getModel();
-		assertEquals(0, model.indexOf(null));
+		FilterComboBoxModel<Item<Integer>> model = (FilterComboBoxModel<Item<Integer>>) comboBox.getModel();
+		assertEquals(0, model.items().visible().indexOf(Item.item(null)));
 		assertTrue(model.items().contains(Item.item(null)));
 
 		assertNull(editModel.value(Detail.INT_VALUE_LIST).get());
@@ -179,7 +178,7 @@ public final class EntityComponentsTest {
 						entityComponents.itemComboBox(Detail.INT_VALUE_LIST)
 										.sorted(false)
 										.buildValue();
-		ItemComboBoxModel<Integer> model = (ItemComboBoxModel<Integer>) componentValue.component().getModel();
+		FilterComboBoxModel<Item<Integer>> model = (FilterComboBoxModel<Item<Integer>>) componentValue.component().getModel();
 
 		//null item should be first, regardless of sorting
 		assertEquals(0, model.items().visible().indexOf(Item.item(null)));